Output ACL files via Moodle instead of performing a redirect
This is mandated by Nextcloud, because Nextcloud increasingly uses
samesite cookies. Therefore, if we redirect requests for files to
Nextcloud in the browser, the browser is not authenticated in Nextcloud.
As a consequence, redirects did not work. This is a workaround that
first loads the file into Moodle (into a per-request directory). The
file is deleted immediately afterwards.